some refactor to use context API
